+++ src/TailwindCSSIntellisensePackage.cs
@@ -34,7 +34,7 @@ namespace TailwindCSSIntellisense
     [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
     [Guid(PackageGuidString)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
-    [ProvideOptionPage(typeof(OptionsProvider.GeneralOptions), "TailwindCSS IntelliSense", "General", 0, 0, true, SupportsProfiles = true)]
+    [ProvideOptionPage(typeof(OptionsProvider.GeneralOptions), "Tailwind CSS IntelliSense", "General", 0, 0, true, SupportsProfiles = true)]
     [InstalledProductRegistration(Vsix.Name, Vsix.Description, Vsix.Version)]
     // Makes sure that the package is initialized early so the solution explorer commands are showing up
     [ProvideAutoLoad(VSConstants.UICONTEXT.NoSolution_string, PackageAutoLoadFlags.BackgroundLoad)]